moved smoke tests to allow creation of functional coverage
