+++ tests/browser/playwright-smoke-tests/page-smoke.spec.js
